GH-1174: extract github/ into internal sub-package
Move GitHub issue management from pkg/orchestrator/issues_gh.go into
pkg/orchestrator/internal/github/issues.go. The parent package retains
thin delegation functions and type aliases for backward compatibility.

Types moved: CobblerIssue, CobblerFrontMatter, ProposedIssue, ContextIssue.
Dependencies injected via Deps struct (Log, GhBin, BranchExists) and
RepoConfig (IssuesRepo, ModulePath, TargetRepo).

Interface at boundary: GitHubIssueManager.
